Added unit test to onlineController - CreateCBFTTransaction

+++ src/main/java/com/mvidyn/std/phub/ui/model/form/payment/ForeignPaymentForm.java
@@ -7,14 +7,14 @@ import lombok.Getter;
 public class ForeignPaymentForm extends PaymentForm {
 
 	private String fxRefNumber;
-	private float exchangeRate;
-	private float creditFxRate;
-	private float debitFxRate;
+	private double exchangeRate;
+	private double creditFxRate;
+	private double debitFxRate;
 
 	@Builder
-	public ForeignPaymentForm(String remittanceCurrency, float remittanceAmount, String paymentCurrency,
-			float paymentAmount, float localEquivalentAmount, String fxRefNumber, float exchangeRate,
-			float creditFxRate, float debitFxRate) {
+	public ForeignPaymentForm(String remittanceCurrency, double remittanceAmount, String paymentCurrency,
+			double paymentAmount, double localEquivalentAmount, String fxRefNumber, double exchangeRate,
+			double creditFxRate, double debitFxRate) {
 		super(remittanceCurrency, remittanceAmount, paymentCurrency, paymentAmount, localEquivalentAmount);
 		this.fxRefNumber = fxRefNumber;
 		this.exchangeRate = exchangeRate;
